+++ R/stan_trans_params.R
@@ -1,65 +1,46 @@
 
-stan_trans_params <- function(prior, lvl_obj) {
+stan_trans_params <- function(prior, lvl_obj, unk_inits) {
 
   var_decl <- paste(
     "  array[n_obs] vector[n_difeq] x; // Output from the ODE solver",
     "  array[n_params] real params;", sep = "\n")
 
-  if(3 == 2) {
+  if(unk_inits) {
 
     var_decl <- paste(var_decl,
                       "  vector[n_difeq] x0; // init values", sep = "\n")
   }
 
+  par_trans_list <- extract_par_trans(prior) %>% remove_NULL()
+
+  if(length(par_trans_list) > 0) {
+
+    par_trans_decl <- sapply(par_trans_list, function(pt_obj) pt_obj$decl) %>%
+      paste(collapse = "\n")
+
+    var_decl <- paste(var_decl, par_trans_decl, sep = "\n")
+  }
+
   # Assignments
-  stock_init_lines <- construct_stock_init_lines(lvl_obj)
-  pars_asg         <- construct_pars_asg(prior)
-  run_model_line   <- "  x = ode_rk45(X_model, x0, t0, ts, params);"
-  asg              <- paste(stock_init_lines, pars_asg, run_model_line,
-                            sep = "\n")
 
-  # unk_elems <- names(unk_list)
-  #
-  # if("measurement" %in% unk_elems) {
-  #
-  #   meas_obj <- lapply(unk_list$measurement, function(par_obj) {
-  #
-  #     obj_elems <- names(par_obj)
-  #
-  #     if("par_trans" %in% obj_elems) {
-  #
-  #       decl  <- stringr::str_glue("  real {par_obj$name};")
-  #       asnmt <- get_trans_line(par_obj)
-  #
-  #
-  #       return( list(decl  = decl,
-  #                    asnmt = asnmt))
-  #     }
-  #
-  #     NULL
-  #
-  #   }) %>% remove_NULL()
-  #
-  # }
-  #
-  #
-  #
+  pars_asg <- construct_pars_asg(prior)
+  asg      <- pars_asg
 
+  if(unk_inits) {
+    stock_init_lines <- construct_stock_init_lines(lvl_obj)
+    asg              <- paste(stock_init_lines, asg, sep = "\n")
+  }
+
+  if(length(par_trans_list) > 0) {
+    par_trans_asg <- sapply(par_trans_list, function(pt_obj) pt_obj$trans) %>%
+      paste(collapse = "\n")
+    asg              <- paste(par_trans_asg, asg, sep = "\n")
+  }
+
+  run_model_line <- "  x = ode_rk45(X_model, x0, t0, ts, params);"
+  asg            <- paste(asg, run_model_line, sep = "\n")
 
   # assignments <- "  //assignments"
-  #
-  # if(length(meas_obj) > 0) {
-  #
-  #   meas_decl_lines <- purrr::map_chr(meas_obj, "decl") %>% paste(sep = "\n")
-  #   var_decl        <- paste(var_decl, meas_decl_lines, sep = "\n")
-  #
-  #   asnmt_lines <- purrr::map_chr(meas_obj, "asnmt") %>% paste(sep = "\n")
-  #   assignments <- paste(assignments, asnmt_lines, sep = "\n")
-  # }
-  #
-  # stock_inits <- construct_stock_init_lines(stock_list, unk_list)
-  # assignments <- paste(assignments, stock_inits, sep = "\n")
-  #
   # net_flow_line <- "  array[n_obs] real delta_x;" # this is optional
   #
   # block_body <- paste(var_decl,
@@ -97,7 +78,7 @@ construct_pars_asg <- function(prior) {
 
   par_names <- sapply(prior, function(prior_obj) {
 
-    if(prior_obj$type == "init") return (NULL)
+    if(prior_obj$type == "init" | prior_obj$type == "meas_par") return (NULL)
 
     prior_obj$par_name
   }) %>%
@@ -107,4 +88,38 @@ construct_pars_asg <- function(prior) {
   stringr::str_glue("  params[{seq_along(par_names)}] = {par_names};") %>%
     paste(collapse = "\n")
 
+}
+
+extract_par_trans <- function(prior) {
+
+  types   <- sapply(prior, function(prior_obj) prior_obj$type)
+  indexes <- which(types == "meas_par")
+
+  meas_pars <- prior[indexes]
+
+  lapply(meas_pars, function(par_obj) {
+
+    obj_elems <- names(par_obj)
+
+    if(!"par_trans" %in% obj_elems) return (NULL)
+
+    pattern <- stringr::str_glue("{par_obj$par_trans}_")
+
+    var_name <- stringr::str_remove(par_obj$par_name, pattern)
+
+    list(decl  = stringr::str_glue("  real {var_name};") %>% as.character(),
+         trans = trans_par(var_name, par_obj$par_trans))
+
+  })
+}
+
+trans_par <- function(var_name, par_trans) {
+
+  if(par_trans == "inv") {
+    return(stringr::str_glue("  {var_name} = 1 / inv_{var_name};") %>%
+      as.character())
+  }
+
+  msg <- stringr::str_glue("Parameter transformation {par_trans} not supported")
+  stop(msg, call. = FALSE)
 }